now ima uploads some new code in my repo

+++ WsCube Tech/add_tow_number.py
@@ -1,7 +1,7 @@
 # x = 10
 # y = 10
-a = int(input("NUmber:"))
-b = int(input("Number:"))
+a = float(input("NUmber:"))
+b = float(input("Number:"))
 sum = a + b
 print(sum)
 # result = x + y
